Move topic creation/deletion test to BaseAdminIntegrationTest
